Move Table into a new file
